Make WithMetadataFrom for open generics public
